back/requirements.txt
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.8.1
async-timeout==5.0.1
bech32==1.2.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
Django==5.1.4
django-admin-sortable2==2.2.4
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-multiselectfield==0.1.13
django-rosetta==0.10.1
django-rq==3.0.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
messagebird==2.2.0
mysqlclient==2.2.6
phonenumbers==8.13.54
pillow==11.0.0
polib==1.2.0
prometheus_client==0.21.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
rq==2.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
